+++ src/SimpleMetaGraphs/src/indsnetwork.jl
