--- src/SubIndexing/src/Project.toml
@@ -1,2 +1,0 @@
-[deps]
-IndexedTables = "6deec6e2-d858-57c5-ab9b-e6ca5bd20e43"
